[FIX] web: `btn-group`, allow gap control
Before this commit, due to the milk design, there was a gap between the
buttons in btn-group. However, in certain modules, it was necessary to
have a negative margin-left to achieve an edge-to-edge appearance.

To address this issue, I introduced a CSS variable called
--btn-group-gap to control the margin-left property for the elements.
This change allows for customization of the margin-left value, with a
fallback value of $btn-border-width by default, when needed.

task-3336458
part of task-3326263

closes odoo/odoo#122126

Related: odoo/enterprise#41357
Signed-off-by: Pierre Paridans (app) <app@odoo.com>